streaming_service/requirements.txt
fastapi
uvicorn
requests
mysql-connector-python 
python-multipart
